Consolidate the aspen.simplates submodule
