+++ aspen/__init__.py
@@ -64,13 +64,6 @@ from __future__ import unicode_literals
 import sys
 import pkg_resources
 
-# imports of convenience
-from . import json_ as json
-from .renderers import BUILTIN_RENDERERS, RENDERERS
-
-# Shut up, PyFlakes. I know I'm addicted to you.
-json, BUILTIN_RENDERERS, RENDERERS
-
 dist = pkg_resources.get_distribution('aspen')
 __version__ = dist.version
 WINDOWS = sys.platform[:3] == 'win'
