+++ aspen/simplates/renderers/json_dump.py
@@ -8,7 +8,7 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 from . import Renderer, Factory
-from .. import json
+from .. import json_
 
 class Renderer(Renderer):
     def compile(self, filepath, raw):
@@ -18,7 +18,7 @@ class Renderer(Renderer):
         output = context['output']
         if not output.media_type:
             output.media_type = context['request_processor'].media_type_json
-        return json.dumps(eval(self.compiled, globals(), context))
+        return json_.dumps(eval(self.compiled, globals(), context))
 
 
 class Factory(Factory):
