+++ aspen/simplates/renderers/stdlib_template.py
