+++ tests/test_dynamic_resource.py
@@ -8,10 +8,10 @@ from pytest import raises, yield_fixture
 from aspen import resources
 from aspen.http.resource import Dynamic
 from aspen.output import Output
-from aspen.pagination import Page
 from aspen.request_processor import dispatcher
-from aspen.renderers.stdlib_template import Factory as TemplateFactory
-from aspen.renderers.stdlib_percent import Factory as PercentFactory
+from aspen.simplates.pagination import Page
+from aspen.simplates.renderers.stdlib_template import Factory as TemplateFactory
+from aspen.simplates.renderers.stdlib_percent import Factory as PercentFactory
 
 
 @yield_fixture
@@ -218,7 +218,7 @@ def test_render_negotation_failures_include_available_types(harness):
     assert actual == expected
 
 
-from aspen.renderers import Renderer, Factory
+from aspen.simplates.renderers import Renderer, Factory
 
 class Glubber(Renderer):
     def render_content(self, context):
