+++ tests/test_renderers.py
@@ -3,8 +3,8 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-from aspen import json
-from aspen.renderers import Factory, Renderer
+from aspen.simplates import json_
+from aspen.simplates.renderers import Factory, Renderer
 
 
 def test_a_custom_renderer(harness):
@@ -15,7 +15,7 @@ def test_a_custom_renderer(harness):
 
         def render_content(self, context):
             d = dict((k, v) for k, v in self.__dict__.items() if k[0] != '_')
-            return json.dumps(d)
+            return json_.dumps(d)
 
     class TestFactory(Factory):
         Renderer = TestRenderer
@@ -27,7 +27,7 @@ def test_a_custom_renderer(harness):
     request_processor.renderer_factories['lorem'] = TestFactory(request_processor)
 
     r = harness.simple("[---]\n[---] text/html via lorem\nLorem ipsum")
-    d = json.loads(r.body)
+    d = json_.loads(r.body)
     assert d['meta'] == 'foobar'
     assert d['raw'] == 'Lorem ipsum'
     assert d['media_type'] == 'text/html'
